requirements.txt
pymupdf
python-docx
scikit-learn
nltk
spacy
